Moving Data and Models to Uplift.DataAccess & Uplift.Models respectively (N-TIER STRUCTURE)
We moved the Models we had in the original Models folder to Uplift.Models

We moved the Data folder (containing migrations and ApplicationDbContext class to Uplift.DataAccess.

We moved the Migrations folder to the root of the Uplift.DataAccess project.

We created two new folders under Uplift.DataAccess > Data (Initializer Repository)

Added dependency references in Uplift main project to Uplift.DataAccess

Added dependency references in Uplift.DataAccess to Uplift.Models and Uplift.Utility